pyproject.toml
[project]
name = "wavelet-loss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pywavelets>=1.8.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
visualize = [
    "matplotlib>=3.10.3",
]
image = [
    "opencv-python>=4.7.0",
    "pillow>=10.0.0",
]
vae = [
    "accelerate>=1.8.1",
    "diffusers>=0.34.0",
]

[tool.ruff]
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wavelet_loss", "src/wavelet_transform"]

[dependency-groups]
dev = [
    "coverage>=7.9.2",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.3",
]
